[생성] filterGymsByDistance 로직 utils 분리
