+++ src/utils/index.js
@@ -14,3 +14,4 @@ export * from './deg2rad.js';
 export * from './getDistanceFromLatLonInKm.js';
 export * from './loadPostcodeScript.js';
 export * from './extractDistrict.js';
+export * from './filterGymsByDistance.js';
